Move export asserts to base class
