+++ test/integration/ggrc/__init__.py
@@ -40,6 +40,22 @@ class TestCase(BaseTestCase, object):
 
   CSV_DIR = os.path.join(THIS_ABS_PATH, "test_csvs/")
 
+  model = None
+
+  DEFAULT_DATETIME_FORMATS = [
+      "{year}",
+      "{year}-{month}",
+      "{year}-{month}-{day}",
+      "{year}-{month}-{day} {hour}",
+      "{year}-{month}-{day} {hour}:{minute}",
+      "{year}-{month}-{day} {hour}:{minute}:{second}",
+      "{month}/{year}",
+      "{month}/{day}/{year}",
+      "{month}/{day}/{year} {hour}",
+      "{month}/{day}/{year} {hour}:{minute}",
+      "{month}/{day}/{year} {hour}:{minute}:{second}",
+  ]
+
   maxDiff = None
 
   @classmethod
@@ -238,3 +254,33 @@ class TestCase(BaseTestCase, object):
         continue
       results[object_type].append(dict(zip(keys, columns[1:])))
     return results
+
+  # pylint: disable=invalid-name
+  def assertSlugs(self, field, value, slugs):
+    """Assert slugs for selected search"""
+    assert self.model
+    search_request = [{
+        "object_name": self.model.__name__,
+        "filters": {
+            "expression": {
+                "left": field,
+                "op": {"name": "="},
+                "right": value,
+            },
+        },
+        "fields": ["slug"],
+    }]
+    parsed_data = self.export_parsed_csv(
+        search_request
+    )[self.model._inflector.title_singular.title()]
+    self.assertEqual(sorted(slugs),
+                     sorted([i["Code*"] for i in parsed_data]))
+
+  # pylint: disable=invalid-name
+  def assertFilterByDatetime(self, alias, datetime_value, slugs, formats=None):
+    """Assert slugs for each date format ent datetime"""
+    parts = ["year", "month", "day", "hour", "minute", "second"]
+    kwargs = {i: getattr(datetime_value, i) for i in parts}
+    formats = formats if formats is not None else self.DEFAULT_DATETIME_FORMATS
+    for f_str in formats:
+      self.assertSlugs(alias, f_str.format(**kwargs), slugs)
